+++ frontend/src/features/analiseEmAndamento/validacao/gerenteTecnico/validacao-analise-tecnica.controller.js
@@ -6,17 +6,16 @@ var ValidacaoAnaliseTecnicaGerenteController = function($rootScope,
                                                     mensagem, 
                                                     $location,
                                                     documentoAnaliseService, 
-                                                    $anchorScroll,
-                                                    processoService, 
-                                                    $uibModal, 
-                                                    empreendimentoService, 
+                                                    $anchorScroll, 
+                                                    $uibModal,      
                                                     documentoService,
                                                     validacaoAnaliseGerenteService, 
                                                     analistaService) {
 
-
     var validacaoAnaliseTecnicaGerente = this;
 
+    validacaoAnaliseTecnicaGerente.analistasTecnicos = null;
+    validacaoAnaliseTecnicaGerente.analistaTecnicoDestino = {};
     validacaoAnaliseTecnicaGerente.analiseTecnicaValidacao = {};
     validacaoAnaliseTecnicaGerente.dadosProjeto = {};
     validacaoAnaliseTecnicaGerente.init = init;
@@ -26,6 +25,7 @@ var ValidacaoAnaliseTecnicaGerenteController = function($rootScope,
     validacaoAnaliseTecnicaGerente.parecerTecnico = {};
     validacaoAnaliseTecnicaGerente.labelDadosProjeto = '';
     validacaoAnaliseTecnicaGerente.enumDocumentos = app.utils.TiposDocumentosAnalise;
+    validacaoAnaliseTecnicaGerente.concluir = concluir;
 
     validacaoAnaliseTecnicaGerente.errors = {
 		despacho: false,
@@ -143,7 +143,7 @@ validacaoAnaliseTecnicaGerente.disable = {
     validacaoAnaliseTecnicaGerente.buscarAnalistasTecnicoByIdProcesso = function() {
 		analistaService.buscarAnalistasTecnicoByIdProcesso(validacaoAnaliseTecnicaGerente.analiseTecnica.analise.processo.id)
 			.then(function(response) {
-				validacaoAnaliseTecnicaGerente.analistaTecnicos = response.data;
+				validacaoAnaliseTecnicaGerente.analistasTecnicos = response.data;
 			});
     };
     
@@ -151,7 +151,7 @@ validacaoAnaliseTecnicaGerente.disable = {
 		$anchorScroll();
 	}
 
-    validacaoAnaliseTecnicaGerente.cancelar =function() {
+    validacaoAnaliseTecnicaGerente.cancelar = function() {
 
         $location.path("/analise-gerente");
     };
@@ -184,7 +184,7 @@ validacaoAnaliseTecnicaGerente.disable = {
 			$('.nav-tabs > .active').next('li').find('a').trigger('click');
 			validacaoAnaliseTecnicaGerente.controleVisualizacao = "ETAPA_VALIDACAO_ANALISE_TECNICA";
 			scrollTop();
-    }, 0);
+        }, 0);
     };
 
     validacaoAnaliseTecnicaGerente.validacaoAbaAvancar = function() {
@@ -335,6 +335,74 @@ validacaoAnaliseTecnicaGerente.disable = {
         });
     };
 
+    function analiseValida(analiseTecnica) {
+
+        if(analiseTecnica.tipoResultadoValidacaoGerente === null || analiseTecnica.tipoResultadoValidacaoGerente === undefined) {
+
+            validacaoAnaliseTecnicaGerente.errors.resultadoAnalise = true;
+            mensagem.error("Preencha os campos obrigatórios para prosseguir com a análise.");
+
+        }else{
+
+            validacaoAnaliseTecnicaGerente.errors.resultadoAnalise = false;
+
+        }
+        
+        if(analiseTecnica.parecerValidacaoGerente === "" || analiseTecnica.parecerValidacaoGerente === null || analiseTecnica.parecerValidacaoGerente === undefined) {
+            
+            validacaoAnaliseTecnicaGerente.errors.despacho = true;
+            mensagem.error("Preencha os campos obrigatórios para prosseguir com a análise.");
+
+        }else{
+
+            validacaoAnaliseTecnicaGerente.errors.resultadoAnalise = false;
+            
+        }
+
+        if(analiseTecnica.tipoResultadoValidacaoGerente.id === validacaoAnaliseTecnicaGerente.TiposResultadoAnalise.PARECER_NAO_VALIDADO.toString() && (validacaoAnaliseTecnicaGerente.analistaTecnicoDestino.id === null || validacaoAnaliseTecnicaGerente.analistaTecnicoDestino.id === undefined)) {
+            
+            validacaoAnaliseTecnicaGerente.errors.analistas = true;
+            mensagem.error("Preencha os campos obrigatórios para prosseguir com a análise.");
+
+        }else{
+
+            validacaoAnaliseTecnicaGerente.errors.analistas = false;
+
+        }
+
+        if(validacaoAnaliseTecnicaGerente.errors.resultadoAnalise === true || validacaoAnaliseTecnicaGerente.errors.despacho === true || validacaoAnaliseTecnicaGerente.errors.analistas === true){
+            return false;
+        }
+        
+        return true;
+
+    }
+
+    function concluir() {
+
+        if(!analiseValida(validacaoAnaliseTecnicaGerente.analiseTecnica)){
+            return;
+        }
+
+        var params = {
+            analiseTecnica: {
+                id: validacaoAnaliseTecnicaGerente.analiseTecnica.id,
+                idAnalistaDestino: validacaoAnaliseTecnicaGerente.analistaTecnicoDestino.id
+            },
+            parecer: validacaoAnaliseTecnicaGerente.analiseTecnica.parecerValidacaoGerente,
+            tipoResultadoAnalise: {id: validacaoAnaliseTecnicaGerente.analiseTecnica.tipoResultadoValidacaoGerente.id}
+        };
+
+        validacaoAnaliseGerenteService.concluirParecerTecnico(params)
+			.then(function(response){
+                $location.path("analise-gerente");
+                $timeout(function() {
+                    mensagem.success("Analise Gerente finalizada!", {referenceId: 5});
+                }, 0);
+            },function(error){
+				mensagem.error(error.data.texto);
+			});
+    }
 };
 
 exports.controllers.ValidacaoAnaliseTecnicaGerenteController = ValidacaoAnaliseTecnicaGerenteController;