+++ CHANGELOG.md
@@ -48,3 +48,4 @@
 
 * OS50 - US213 - [Gerente Técnico] Eu, gerente técnico, devo acessar o processo para validar análise técnica.
 * OS50 - US1398 - [Gerente Técnico] Eu, gerente técnico, devo acessar a análise técnica para validação.
+* OS50 - US1415 - [Gerente Técnico] Eu, gerente técnico, devo concluir a análise para validar a aba conclusão da análise